+++ examples/demo3/src/index/App.vue
@@ -4,7 +4,7 @@
       <div class="label">{{item}}</div>
       <div class="comp">
         <!-- 可使用 html 标签替代的内置组件 -->
-        <img v-if="item === 'image'" src="https://res.wx.qq.com/wxdoc/dist/assets/img/0.4cb08bb4.jpg" width="50" height="50" />
+        <img v-if="item === 'img'" src="https://res.wx.qq.com/wxdoc/dist/assets/img/0.4cb08bb4.jpg" width="50" height="50" @load="onImgLoad" />
         <input v-else-if="item === 'input'" type="number" placeholder="请输入内容" @input="onInput" />
         <textarea v-else-if="item === 'textarea'" placeholder="请输入内容" maxlength="50" :auto-height="true" value="我是 textarea" @input="onTextareaInput" />
         <video class="video" v-else-if="item === 'video'" src="http://wxsnsdy.tc.qq.com/105/20210/snsdyvideodownload?filekey=30280201010421301f0201690402534804102ca905ce620b1241b726bc41dcff44e00204012882540400&bizid=1023&hy=SH&fileparam=302c020101042530230204136ffd93020457e3c4ff02024ef202031e8d7f02030f42400204045a320a0201000400" :muted="true" :show-mute-btn="true" :controls="true">
@@ -26,6 +26,7 @@
         <!-- 使用 wx-component 来创建内置组件 -->
         <wx-component v-else-if="item === 'view'" behavior="view">我是视图</wx-component>
         <wx-component v-else-if="item === 'button'" behavior="button" open-type="share">分享</wx-component>
+        <wx-component v-else-if="item === 'image'" behavior="image" src="https://res.wx.qq.com/wxdoc/dist/assets/img/0.4cb08bb4.jpg"></wx-component>
         <div v-else-if="item === 'picker'">
           <wx-component behavior="picker" :value="1" :range="['美国', '中国', '巴西', '日本']">点击&nbsp;&nbsp;选择国家</wx-component>
           <wx-component behavior="picker" mode="region" @change="onPickerChange">
@@ -34,7 +35,10 @@
           </wx-component>
         </div>
         <wx-component v-else-if="item === 'map'" behavior="map" class="map" :longitude="113.324520" :latitude="23.099994" :scale="14" :controls="map.controls" :markers="map.markers" :polyline="map.polyline" :show-location="true" @markertap="onMapMarkerTap" @regionchange="onMapRegionChange" @controltap="onMapControlTap"></wx-component>
-        <wx-compoennt v-else-if="item === 'cover-view'">测试 cover-view</wx-compoennt>
+        <wx-compoennt v-else-if="item === 'cover-view'" behavior="cover-view">测试 cover-view</wx-compoennt>
+        <wx-component v-else-if="item === 'cover-image'" behavior="cover-view">
+          <wx-component behavior="cover-image" src="https://res.wx.qq.com/wxdoc/dist/assets/img/0.4cb08bb4.jpg"></wx-component>
+        </wx-component>
         <wx-component v-else-if="item === 'live-player'" behavior="live-player" class="live-player" mode="live" :autoplay="true" src="rtmp://live.hkstv.hk.lxdns.com/live/hks" @statechange="onLivePlayerStateChange"></wx-component>
         <wx-component v-else-if="item === 'live-pusher'" behavior="live-pusher" class="live-pusher" mode="RTC" :autopush="true" url="https://domain/push_stream" @statechange="onLivePusherStateChange"></wx-component>
         <!-- 不支持标签 -->
@@ -50,15 +54,17 @@ export default {
   data() {
     return {
       list: [
-        'image',
+        'img',
         'input',
         'textarea',
         'video',
         'view',
         'button',
         'picker',
+        'image',
         'map',
         'cover-view',
+        'cover-image',
         'live-player',
         'live-pusher',
         'iframe',
@@ -110,6 +116,10 @@ export default {
 
     onTextareaInput(evt) {
       console.log('onTextareaInput', evt.target.value)
+    },
+
+    onImgLoad(evt) {
+      console.log('onImgLoad', evt.detail)
     },
 
     onPickerChange(evt) {