+++ packages/miniprogram-element/src/index.js
@@ -217,6 +217,18 @@ Component({
             }
         },
 
+        onImgLoad(evt) {
+            if (this.document.$$checkEvent(evt)) {
+                this.callEvent(evt, 'load', {width: evt.detail.width, height: evt.detail.height})
+            }
+        },
+
+        onImgError(evt) {
+            if (this.document.$$checkEvent(evt)) {
+                this.callEvent(evt, 'error')
+            }
+        },
+
         ...initHandle,
     }
 })