+++ packages/miniprogram-element/src/util/init-data.js
@@ -7,6 +7,7 @@ const picker = require('../component/picker')
 const button = require('../component/button')
 const map = require('../component/map')
 const coverView = require('../component/cover-view')
+const coverImage = require('../component/cover-image')
 const livePlayer = require('../component/live-player')
 const livePusher = require('../component/live-pusher')
 
@@ -20,6 +21,7 @@ module.exports = {
     button: button.properties,
     map: map.properties,
     'cover-view': coverView.properties,
+    'cover-image': coverImage.properties,
     'live-player': livePlayer.properties,
     'live-pusher': livePusher.properties,
 }