+++ packages/miniprogram-element/src/util/init-handle.js
@@ -9,6 +9,7 @@ const picker = require('../component/picker')
 const button = require('../component/button')
 const map = require('../component/map')
 const coverView = require('../component/cover-view')
+const coverImage = require('../component/cover-image')
 const livePlayer = require('../component/live-player')
 const livePusher = require('../component/live-pusher')
 
@@ -70,6 +71,7 @@ module.exports = {
     ...button.handles,
     ...map.handles,
     ...coverView.handles,
+    ...coverImage.handles,
     ...livePlayer.handles,
     ...livePusher.handles,
 }